+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/TenantResponseDto.java
@@ -41,6 +41,17 @@ public class TenantResponseDto {
     @Valid
     public Boolean node;
 
+
+    @Schema(
+            type = SchemaType.BOOLEAN,
+            implementation = Boolean.class,
+            description = "Indicates whether the tenant is configured to have performance data",
+            example = "true"
+    )
+    @JsonProperty("performance")
+    @Valid
+    public Boolean performance;
+
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,